own_project/pyproject.toml
[tool.poetry]
name = "new_admin_panel_sprint_1"
version = "0.1.0"
description = ""
authors = ["Mikhail <golobokov.misha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
python-dotenv = "^0.20.0"
django-split-settings = "^1.1.0"
uWSGI = "^2.0.20"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.13.0"
pydantic = "^1.9.1"
django-extensions = "^3.2.0"
django-debug-toolbar = "3.4.0"

[tool.poetry.dev-dependencies]
flake8 = "4.0"
Faker = "^13.14.0"
pre-commit = "^2.19.0"
more-itertools = "^8.13.0"
pytest = "^7.1.2"
django-debug-toolbar = "3.4.0"
django-extensions = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
